sakai-10.0

+++ kernel/kernel-impl/src/main/java/org/sakaiproject/user/impl/AuthnCacheWatcher.java
@@ -1,6 +1,6 @@
 /**********************************************************************************
- * $URL: https://source.sakaiproject.org/svn/kernel/tags/kernel-1.3.3/kernel-impl/src/main/java/org/sakaiproject/user/impl/AuthnCacheWatcher.java $
- * $Id: AuthnCacheWatcher.java 81062 2010-08-11 14:39:44Z david.horwitz@uct.ac.za $
+ * $URL: https://source.sakaiproject.org/svn/kernel/tags/sakai-10.0/kernel-impl/src/main/java/org/sakaiproject/user/impl/AuthnCacheWatcher.java $
+ * $Id: AuthnCacheWatcher.java 309199 2014-05-06 15:36:14Z enietzel@anisakai.com $
  ***********************************************************************************
  *
  * Copyright (c) 2005, 2006, 2007, 2008 Sakai Foundation
@@ -9,7 +9,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *       http://www.osedu.org/licenses/ECL-2.0
+ *       http://www.opensource.org/licenses/ECL-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -21,19 +21,20 @@
 
 package org.sakaiproject.user.impl;
 
-import java.util.Observable;
-import java.util.Observer;
-
-import net.sf.ehcache.Cache;
-
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.sakaiproject.component.cover.ComponentManager;
 import org.sakaiproject.entity.api.EntityManager;
 import org.sakaiproject.entity.api.Reference;
 import org.sakaiproject.event.api.Event;
 import org.sakaiproject.event.api.EventTrackingService;
+import org.sakaiproject.memory.api.Cache;
+import org.sakaiproject.memory.api.MemoryService;
 import org.sakaiproject.user.api.UserDirectoryService;
 import org.sakaiproject.user.api.UserNotDefinedException;
+
+import java.util.Observable;
+import java.util.Observer;
 
 /**
  * This observer watches for user.add and user.upd events to invalidate the Authn cache
@@ -44,20 +45,20 @@ import org.sakaiproject.user.api.UserNotDefinedException;
 public class AuthnCacheWatcher implements Observer {
 
 	private static final Log log = LogFactory.getLog(AuthnCacheWatcher.class);
-	
-	private AuthenticationCache authenticationCache;
-	
-	private UserDirectoryService userDirectoryService;
-	
-	private EventTrackingService eventTrackingService;
-	
-	private EntityManager entityManager;
-	
 	//Copied from DbUserService as they are private
 	private static final String EIDCACHE = "eid:";
 	private static final String IDCACHE = "id:";
-	
+	private AuthenticationCache authenticationCache;
+	private UserDirectoryService userDirectoryService;
+	private EventTrackingService eventTrackingService;
+	private EntityManager entityManager;
+    private MemoryService memoryService;
 	private Cache userCache = null;
+	
+    public void setMemoryService(MemoryService memoryService) {
+        this.memoryService = memoryService;
+    }
+
 		public void setUserCache(Cache userCache) {
 		this.userCache = userCache;
 	}
@@ -85,11 +86,16 @@ public class AuthnCacheWatcher implements Observer {
 
 	public void init() {
 		log.info("init()");
+        if (userCache == null) { // this is the user id->eid mapping cache
+            userCache = memoryService.getCache("org.sakaiproject.user.api.UserDirectoryService");
+        }
 		eventTrackingService.addObserver(this);
 	}
 	
 	public void destroy() {
-		eventTrackingService.deleteObserver(this);
+        if (!ComponentManager.hasBeenClosed()) {
+            eventTrackingService.deleteObserver(this);
+        }
 	}
 	
 	public void update(Observable arg0, Object arg) {
